Add files via upload
The program begins by defining a permutation array, which represents the desired order of elements. In this example, the permutation array is initialized with the values [2, 0, 1, 4, 3].

Next, a new array called "array" is created with the same length as the permutation array. This array will be used to store the elements in the desired order.

A for loop is then used to iterate over each index of the permutation array. During each iteration, the element at the current index of the permutation array is assigned to the corresponding index in the array array.

Finally, the permutation array and the resulting array are printed using the Arrays.toString() method for better readability.

The output of the program shows the original permutation and the resulting array, which should be the same in this case:

Permutation: [2, 0, 1, 4, 3]
Array: [2, 0, 1, 4, 3]

This code provides a basic example of how to construct an array from a given permutation, allowing you to work with the elements in the desired order.