Add files via upload
The Transpose class is defined within the com.ritesh package.

The program creates a 2x2 integer array called a to store the entered data.

The code prompts the user to enter array data by displaying the message "Enter Array Data:".

It creates a Scanner object called in to read input from the user.

Using nested for loops, the program iterates over each element of the 2x2 array and assigns the user-entered values to the corresponding elements.

After reading the array data, the program prints a new line ("\n") to provide separation between the input and the output.

Another set of nested for loops is used to iterate over the array elements and print them. Each element is printed followed by a space.

After printing each row of the array, a new line is printed to display the matrix form.

In summary, this program allows the user to input data into a 2x2 array and then displays the array in matrix form.